django_project/requirements.txt
asgiref==3.7.2
Django==4.2.1
django-widget-tweaks==1.4.12
sqlparse==0.4.4
tzdata==2023.3
mysql-connector-python
